fix bugs FE and update modal product
