+++ projects/server/src/routers/stockMutationRouter.js
@@ -13,6 +13,6 @@ const { verifyRoleAdmin, verifyToken } = require("../middleware/verifyToken");
 Router.post("/request-stock", verifyRoleAdmin, validateRequestStock, stockMutation.requestStock);
 Router.patch("/confirm-mutation/:id", verifyToken, stockMutation.confirmRequest);
 Router.get("/getrequest-stock", verifyToken, stockMutation.getStockRequest);
-Router.get("/getAllRequestMutation", stockMutation.getAllRequestMutation);
+Router.get("/getAllRequestMutation", verifyRoleAdmin, stockMutation.getAllRequestMutation);
 
 module.exports = Router;
